sessionの管理を外だし。
